refactor: extract clipboard, auto-copy, and console theme from index.tsx
Move clipboard management, auto-copy-on-select behavior, and console
theme syncing into dedicated modules. Fixes infinite render loop caused
by calling requestRender() inside a post-process function.
